edit and deleting button now popping up
