+++ .env
@@ -1,4 +1,4 @@
 VITE_ENCRYPT_KEY = "RadsDashboard#$%245"
-VITE_BACKEND_BASEURL = "https://staging-backend.rdas.com.ng/api/v1/"
-# VITE_BACKEND_BASEURL =  "http://127.0.0.1:80/api/v1/"
+#VITE_BACKEND_BASEURL = "https://staging-backend.rdas.com.ng/api/v1/"
+VITE_BACKEND_BASEURL =  "http://127.0.0.1:80/api/v1/"
