+++ src/composable/useEditComposable.js
@@ -4,6 +4,8 @@ import { ref} from 'vue';
 import apiService from '@/services/apiService';
 //import { useReadComposable} from '@/composable/useReadComposable';
 import { catchAxiosError, catchAxiosSuccess } from '@/services/Response'
+import { useStore } from "@/stores/user";
+import { computed } from 'vue';
 
 
 //const { fetchPage } = useReadComposable();
@@ -15,6 +17,23 @@ export function useEditComposable(formFields, url,itemId,emit) {
   let items = ref();
   let errorMessage = ref();
   let loading = ref(false);
+
+  const store = useStore();
+const permissions = computed(() => {
+    
+    return  store.getUser.user.permission.permissions.find(p => p.page_name === "measurements");
+ })
+
+const additionalColumns = computed(() => {
+    const cols = [];
+    if (permissions.value?.update) {
+      cols.push({ name: 'Edit', action: handleEdit });
+    }
+    if (permissions.value?.del) {
+      cols.push({ name: 'Delete', action: handleDelete });
+    }
+    return cols;
+  });
   //const forceUpdate = ref(true);
  
 
@@ -25,10 +44,13 @@ export function useEditComposable(formFields, url,itemId,emit) {
   };
   
   const handleEdit = (item='') => {
-
+    console.log('before value')
+    console.log(showEditModal.value)
     console.log(item)
     items.value =item;
     showEditModal.value = true;
+    console.log('after value')
+    console.log(showEditModal.value)
     
   };
 
@@ -63,7 +85,7 @@ export function useEditComposable(formFields, url,itemId,emit) {
 
       console.log(url);
       console.log(itemId);
-      itemId = "9bc64d90-76e8-4f2a-b446-9d7f8e2180f6";
+      //itemId = "9bc64d90-76e8-4f2a-b446-9d7f8e2180f6";
        const Url = `${url}/${itemId}`
       
       
@@ -111,7 +133,7 @@ export function useEditComposable(formFields, url,itemId,emit) {
     items,
     editForm,
     loading,
-    
+    additionalColumns
   
   };
 }
